[Enhancement] support batch get pk index when partial update by column (#31228)
Fixes #20436
Merge multi-segment files' PK at once, and batch get index to reduce read IO amplification. And also control memory usage under config `primary_key_batch_get_index_memory_limit`.

Test Result:
87 segment files generated when data loading, we compare the latency of getting from the index, under different batch get key size.
|      primary_key_batch_get_index_memory_limit size     | Time  |
| --------- | ----- |
| No batch  | 82s   |
| 100MB     | 54s   |
| 1000MB    | 27s   |
| 10000MB   | 10s   |

+++ be/test/storage/rowset_column_partial_update_test.cpp
@@ -40,11 +40,13 @@
 #include "testutil/assert.h"
 
 namespace starrocks {
-class RowsetColumnPartialUpdateTest : public ::testing::Test {
+
+class RowsetColumnPartialUpdateTest : public ::testing::Test, testing::WithParamInterface<int64_t> {
 public:
     void SetUp() override {
         _compaction_mem_tracker = std::make_unique<MemTracker>(-1);
         _update_mem_tracker = std::make_unique<MemTracker>();
+        config::primary_key_batch_get_index_memory_limit = GetParam();
     }
 
     void TearDown() override {
@@ -405,7 +407,7 @@ static void prepare_tablet(RowsetColumnPartialUpdateTest* self, const TabletShar
     }
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, partial_update_and_check) {
+TEST_P(RowsetColumnPartialUpdateTest, partial_update_and_check) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -453,7 +455,7 @@ TEST_F(RowsetColumnPartialUpdateTest, partial_update_and_check) {
     }));
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, normal_partial_update_and_check) {
+TEST_P(RowsetColumnPartialUpdateTest, normal_partial_update_and_check) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -485,7 +487,7 @@ TEST_F(RowsetColumnPartialUpdateTest, normal_partial_update_and_check) {
     }));
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, partial_update_diff_column_and_check) {
+TEST_P(RowsetColumnPartialUpdateTest, partial_update_diff_column_and_check) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -523,7 +525,7 @@ TEST_F(RowsetColumnPartialUpdateTest, partial_update_diff_column_and_check) {
     ASSERT_TRUE(StorageEngine::instance()->update_manager()->TEST_primary_index_refcnt(tablet->tablet_id(), 1));
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, partial_update_multi_segment_and_check) {
+TEST_P(RowsetColumnPartialUpdateTest, partial_update_multi_segment_and_check) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -562,7 +564,7 @@ TEST_F(RowsetColumnPartialUpdateTest, partial_update_multi_segment_and_check) {
     ASSERT_TRUE(StorageEngine::instance()->update_manager()->TEST_primary_index_refcnt(tablet->tablet_id(), 1));
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, partial_update_multi_segment_preload_and_check) {
+TEST_P(RowsetColumnPartialUpdateTest, partial_update_multi_segment_preload_and_check) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -571,7 +573,7 @@ TEST_F(RowsetColumnPartialUpdateTest, partial_update_multi_segment_preload_and_c
     prepare_tablet(this, tablet, version, version_before_partial_update, N);
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, partial_update_compaction_and_check) {
+TEST_P(RowsetColumnPartialUpdateTest, partial_update_compaction_and_check) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -589,7 +591,7 @@ TEST_F(RowsetColumnPartialUpdateTest, partial_update_compaction_and_check) {
     }
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, test_full_clone) {
+TEST_P(RowsetColumnPartialUpdateTest, TEST_Pull_clone) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -599,7 +601,7 @@ TEST_F(RowsetColumnPartialUpdateTest, test_full_clone) {
 
     {
         // clone from _tablet to new_tablet
-        auto new_tablet = create_tablet(2000, 2000);
+        auto new_tablet = create_tablet(rand(), rand());
         ASSERT_EQ(1, new_tablet->updates()->version_history_count());
         ASSERT_OK(full_clone(tablet, version, new_tablet));
         ASSERT_TRUE(check_tablet(new_tablet, version, N, [](int64_t k1, int64_t v1, int32_t v2) {
@@ -609,7 +611,7 @@ TEST_F(RowsetColumnPartialUpdateTest, test_full_clone) {
 
     {
         // clone from _tablet to new_tablet with version before partial update
-        auto new_tablet = create_tablet(3000, 3000);
+        auto new_tablet = create_tablet(rand(), rand());
         ASSERT_EQ(1, new_tablet->updates()->version_history_count());
         ASSERT_OK(full_clone(tablet, version_before_partial_update, new_tablet));
         ASSERT_TRUE(check_tablet(new_tablet, version_before_partial_update, N, [](int64_t k1, int64_t v1, int32_t v2) {
@@ -619,7 +621,7 @@ TEST_F(RowsetColumnPartialUpdateTest, test_full_clone) {
 
     {
         // clone from _tablet to new_tablet with version just after first partial update
-        auto new_tablet = create_tablet(4000, 4000);
+        auto new_tablet = create_tablet(rand(), rand());
         ASSERT_EQ(1, new_tablet->updates()->version_history_count());
         ASSERT_OK(full_clone(tablet, version_before_partial_update + 1, new_tablet));
         ASSERT_TRUE(
@@ -629,7 +631,7 @@ TEST_F(RowsetColumnPartialUpdateTest, test_full_clone) {
     }
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, test_increment_clone) {
+TEST_P(RowsetColumnPartialUpdateTest, test_increment_clone) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -639,7 +641,7 @@ TEST_F(RowsetColumnPartialUpdateTest, test_increment_clone) {
 
     {
         // 1. full clone with version before partial update
-        auto new_tablet = create_tablet(5000, 5000);
+        auto new_tablet = create_tablet(rand(), rand());
         ASSERT_EQ(1, new_tablet->updates()->version_history_count());
         ASSERT_OK(full_clone(tablet, version_before_partial_update, new_tablet));
         ASSERT_TRUE(check_tablet(new_tablet, version_before_partial_update, N, [](int64_t k1, int64_t v1, int32_t v2) {
@@ -660,7 +662,7 @@ TEST_F(RowsetColumnPartialUpdateTest, test_increment_clone) {
     }
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, test_schema_change) {
+TEST_P(RowsetColumnPartialUpdateTest, test_schema_change) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -670,7 +672,7 @@ TEST_F(RowsetColumnPartialUpdateTest, test_schema_change) {
 
     {
         // create table with add column, test link_from
-        auto new_tablet = create_tablet(6000, 6000, true);
+        auto new_tablet = create_tablet(rand(), rand(), true);
         new_tablet->set_tablet_state(TABLET_NOTREADY);
         auto chunk_changer = std::make_unique<ChunkChanger>(new_tablet->tablet_schema());
         ASSERT_TRUE(new_tablet->updates()->link_from(tablet.get(), version, chunk_changer.get()).ok());
@@ -681,7 +683,7 @@ TEST_F(RowsetColumnPartialUpdateTest, test_schema_change) {
     }
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, test_full_clone2) {
+TEST_P(RowsetColumnPartialUpdateTest, TEST_Pull_clone2) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -702,7 +704,9 @@ TEST_F(RowsetColumnPartialUpdateTest, test_full_clone2) {
     }
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, test_dcg_gc) {
+TEST_P(RowsetColumnPartialUpdateTest, test_dcg_gc) {
+    // Only run one parameter here
+    if (GetParam() != 104857600) return;
     fs::remove_all(get_stores()->path());
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
@@ -759,7 +763,7 @@ TEST_F(RowsetColumnPartialUpdateTest, test_dcg_gc) {
     }));
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, test_get_column_values) {
+TEST_P(RowsetColumnPartialUpdateTest, test_get_column_values) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -790,7 +794,7 @@ TEST_F(RowsetColumnPartialUpdateTest, test_get_column_values) {
     }
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, test_upsert) {
+TEST_P(RowsetColumnPartialUpdateTest, test_upsert) {
     const int N = 100;
     auto tablet = create_tablet(rand(), rand());
     ASSERT_EQ(1, tablet->updates()->version_history_count());
@@ -833,7 +837,7 @@ TEST_F(RowsetColumnPartialUpdateTest, test_upsert) {
     }
 }
 
-TEST_F(RowsetColumnPartialUpdateTest, partial_update_two_rowset_and_check) {
+TEST_P(RowsetColumnPartialUpdateTest, partial_update_two_rowset_and_check) {
     const int N = 100;
     const int M = N / 2;
     auto tablet = create_tablet(rand(), rand());
@@ -887,5 +891,8 @@ TEST_F(RowsetColumnPartialUpdateTest, partial_update_two_rowset_and_check) {
         return (int16_t)(k1 % 100 + 3) == v1 && (int32_t)(k1 % 1000 + 2) == v2;
     }));
 }
+
+INSTANTIATE_TEST_SUITE_P(RowsetColumnPartialUpdateTest, RowsetColumnPartialUpdateTest,
+                         ::testing::Values(1, 1024, 104857600));
 
 } // namespace starrocks
